+++ src/services/node_service.py
@@ -210,14 +210,16 @@ class NodeService:
 
     def update(self, node_id: int, updates: NodeUpdate, include_deleted = False) -> Node:
         node = self.get_node(node_id, include_deleted)
-        print("updates:", updates)
-
-        for field, value in updates:
-            if value is not None:
-                setattr(node, field, value)
+        print("before_update: ", node)
+        print("update: ", updates)
+        
+        # Fields explicitly provided in updates (even if set to None) will overwrite existing values (due to model_fiels_set).
+        # To prevent setting a field to None,y do not include it in the update payload.
+        for field in updates.model_fields_set: 
+            value = getattr(updates, field)
+            setattr(node, field, value)
 
         self._repo.update(node)
-
         return node
         
     def get_due_nodes(self, collection_id: int) -> list[Node]: